Move action CRUD methods to separate route. Clean up comments and console.logs
